refactor(llm-nodes): split agents.ts god-file into flat sibling modules
Pure structural refactor of packages/llm-nodes/src/nodes/agents.ts
(3,436 lines) — no behavior change. Split the shared helpers, data, and
support types out into flat sibling modules, leaving agents.ts as the
seven node classes, the system-prompt constants, and a barrel that
re-exports every symbol it exported before, so all existing importers
compile unchanged.

New modules under src/nodes/:
- agent-utils.ts            asText/summarize/tokenize/extractJson, model
                            + provider helpers, message/tool normalizers,
                            stream helpers, ToolLike, structured-output
                            schema
- agent-threads.ts          THREAD_STORE + thread load/save helpers
- agent-thinking.ts         redacted-thinking constants + stream splitter
- agent-control-tools.ts    control-tool marker, build/isControlTool
- agent-loop.ts             runAgentLoop, AgentLoop{Options,Result},
                            ToolLikeAdapter
- agent-recommended-models.ts  per-node recommendedModels catalogs +
                            GEMMA_3_4B_IT_GGUF_* constants (data)

agents.ts re-exports expandAssetReferences, ToolLike, streamProviderMessages,
isChunkItem, isToolCallItem, toProviderTools, serializeToolResult,
runAgentLoop, AgentLoopOptions, AgentLoopResult, the node classes, and
AGENT_NODES. Intra-package imports use .js extensions.

Hygiene (still zero behavior change):
- drop (this as any).max_tokens casts in Extractor/Classifier
- route logThreadWarning through the module logger (log.warn), keep the
  NODE_ENV==="test" guard (no test depends on console.warn)
- drop the duplicated inputKeys field in AgentNode's start log
